fixed bugs, added pretty colors to tui
